Update to 2.5-alpha6

+++ src/main/java/ch/njol/yggdrasil/JRESerializer.java
@@ -132,7 +132,7 @@ public class JRESerializer extends YggdrasilSerializer<Object> {
 	@Override
 	public <E> E deserialize(Class<E> c, Fields fields) throws StreamCorruptedException, NotSerializableException {
 		if (c == UUID.class) {
-			return (E) new UUID(((Long)fields.getPrimitive("mostSigBits", Long.TYPE)).longValue(), ((Long)fields.getPrimitive("leastSigBits", Long.TYPE)).longValue());
+			return (E) new UUID(fields.getPrimitive("mostSigBits", Long.TYPE).longValue(), fields.getPrimitive("leastSigBits", Long.TYPE).longValue());
 		}
 		throw new StreamCorruptedException();
 	}
